requirements.txt
# GirlBot - Telegram Fitness Assistant
# Python 3.12+

# Telegram Bot
python-telegram-bot[all]>=22.0

# AI/LLM - let langgraph resolve compatible versions
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langsmith>=0.3.0
openai>=1.50.0
anthropic>=0.40.0

# Database
supabase>=2.10.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
pgvector>=0.3.0

# Cache
redis[hiredis]>=5.0.0

# HTTP/Async
httpx>=0.27.0
aiohttp>=3.10.0

# Configuration
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# Resilience
pybreaker>=1.2.0
aiolimiter>=1.1.0
tenacity>=9.0.0

# Logging
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0

# Development
black>=24.10.0
ruff>=0.8.0
mypy>=1.13.0
